Move models to 3d dir
