+++ js/components/HomeScene/DebugModelScene.jsx
@@ -1,6 +1,6 @@
 import { CameraControls, Center, Environment } from "@react-three/drei";
 
-import * as Models from "./Models";
+import * as Models from "../3D/Models";
 
 export default function DebugModelScene({ model = "CareersModel" }) {
     const Component = Models[model];
